Move aggregators to aggregators.js
